+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `phpCodeSniffer.exec.linux`, `phpCodeSniffer.exec.osx`, and `phpCodeSniffer.exec.windows` options
+for platform-specific executables.
+- Support for execution on Windows without the use of WSL.
+
+### Deprecated
+- `phpCodeSniffer.executable` has been deprecated in favor of using platform-specific executables.
 
 ## [1.7.0] - 2022-07-29
 ### Added